requirements.txt
flask
flask-wtf
email_validator
faker
python-dotenv
flask-migrate
wtforms-sqlalchemy
bootstrap-flask
flask-login